Implicitly cast a size_t value to an integer,  in 'calendar' demo and in generic clipboard function.  Errors spotted by MSVC.

+++ common/pdcclip.c
@@ -66,7 +66,7 @@ int PDC_getclipboard(char **contents, long *length)
     if (!pdc_clipboard)
         return PDC_CLIP_EMPTY;
 
-    len = strlen(pdc_clipboard);
+    len = (int)strlen(pdc_clipboard);
     *contents = malloc(len + 1);
     if (!*contents)
         return PDC_CLIP_MEMORY_ERROR;
